# swagger.yaml
openapi: 3.0.0
info:
  title: Multi-Tenant Calendar API
  version: 1.0.0
  description: API for managing tenant events with Google Calendar integration
servers:
  - url: https://dev.matterminer.com/calendar
    description: Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /auth/google:
    get:
      summary: Initialize Google OAuth flow
      description: Redirects the user to Google's consent screen. Requires a tenantId query parameter to track the session.
      tags: [Auth]
      parameters:
        - in: query
          name: tenantId
          schema:
            type: string
          required: true
          description: Unique ID for the tenant (defaults to 'default_tenant')
      responses:
        302:
          description: Redirect to Google Authorization Server
        500:
          description: Internal Server Error
  /auth/oauth2callback:
    get:
      summary: Google OAuth callback handler
      description: This endpoint is called by Google after the user authorizes the app. It exchanges the auth code for tokens, stores them in Redis, and issues a JWT.
      tags: [Auth]
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
          description: The authorization code provided by Google
        - in: query
          name: state
          schema:
            type: string
          description: Usually contains the tenantId passed during the initial request
      responses:
        200:
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  jwtToken: { type: string, description: "Bearer token for subsequent API calls" }
                  tenantId: { type: string }
        400:
          description: Missing authorization code
        500:
          description: Failed to exchange tokens or store session
...